Fix the following bugs (on the description)
The first operand wasn't diplayed when a operator was clicked. Same
for the result when a second or another operator was clicked.

Pressing "=" before entering all of the numbers or an operator caused
problems.

Dividing by 0 is now stopping the operations when an operator is clicked
and clears the display and all the data.